pyproject.toml
[project]
name = "searchmark-api"
version = "0.0.1"
description = "API backend for paid feature to SearchMark that uses an LLM to suggest the best bookmark folder for the current page."
authors = [{ name = "Jonathan Piaget", email = "pro@piagetjonathan.ch" }]
readme = "README.md"
keywords = ['python']
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi[standard]",
    "litellm",
    "html2text",
    "python-toon",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://JonathanPiaget.github.io/searchmark-api/"
Repository = "https://github.com/JonathanPiaget/searchmark-api"
Documentation = "https://JonathanPiaget.github.io/searchmark-api/"

[project.scripts]
searchmark-test = "app.testing.runner:run"

[project.optional-dependencies]
testing = ["rich"]

[dependency-groups]
dev = [
    "pytest",
    "httpx",
    "pre-commit",
    "tox-uv",
    "ty",
    "ruff",
    "rich",
]
prod = ["sentry-sdk"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.10"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"app/testing/*" = ["TRY003", "C901", "S112"]

[tool.ruff.format]
preview = true
